Changes character of randomness
I decided my earlier implementation of randomness was a little too
random. This new algorithm chooses one random number somewhere between 4
and 7 seconds and then makes all the other frequencies just a little out
of sync with that.

The result is more interesting, IMO, if you have the patients to listen
for a few minutes.